fixed circular dependencies and cloud templates
