+++ x-pack/solutions/security/packages/kbn-cloud-security-posture/public/src/components/fleet_extensions/aws_credentials_form/aws_hooks.ts
@@ -7,16 +7,22 @@
 
 import { useEffect, useRef } from 'react';
 import { NewPackagePolicy, NewPackagePolicyInput, PackageInfo } from '@kbn/fleet-plugin/common';
-import { AWS_CREDENTIALS_TYPE, AWS_SETUP_FORMAT } from '../constants';
-import { getPosturePolicy } from '../utils';
 import {
+  AWS_CREDENTIALS_TYPE,
+  AWS_SETUP_FORMAT,
   DEFAULT_MANUAL_AWS_CREDENTIALS_TYPE,
+} from '../constants';
+import {
+  updatePolicyWithInputs,
+  getAwsCredentialsType,
+  getCloudFormationDefaultValue,
+} from '../utils';
+import {
   getAwsCredentialsFormOptions,
   getInputVarsFields,
 } from './get_aws_credentials_form_options';
 import { AwsCredentialsType, AwsSetupFormat, UpdatePolicy } from '../types';
-import { getAwsCredentialsType, getCloudFormationDefaultValue } from './aws_utils';
-import { useCloudSetup } from '../cloud_setup_context';
+import { useCloudSetup } from '../hooks/use_cloud_setup_context';
 /**
  * Update CloudFormation template and stack name in the Agent Policy
  * based on the selected policy template
@@ -76,7 +82,7 @@ export const useAwsCredentialsForm = ({
     // This should ony set the credentials after the initial render
     if (!getAwsCredentialsType(input) && !lastManualCredentialsType.current) {
       updatePolicy({
-        updatedPolicy: getPosturePolicy(newPolicy, awsPolicyType, {
+        updatedPolicy: updatePolicyWithInputs(newPolicy, awsPolicyType, {
           'aws.credentials.type': {
             value: awsCredentialsType,
             type: 'text',
@@ -98,6 +104,8 @@ export const useAwsCredentialsForm = ({
     newPolicy,
     updatePolicy,
     setupFormat,
+    awsPolicyType,
+    templateName,
   });
 
   const onSetupFormatChange = (newSetupFormat: AwsSetupFormat) => {
@@ -110,7 +118,7 @@ export const useAwsCredentialsForm = ({
       lastManualCredentialsType.current = getAwsCredentialsType(input);
 
       updatePolicy({
-        updatedPolicy: getPosturePolicy(newPolicy, awsPolicyType, {
+        updatedPolicy: updatePolicyWithInputs(newPolicy, awsPolicyType, {
           'aws.credentials.type': {
             value: AWS_CREDENTIALS_TYPE.CLOUD_FORMATION,
             type: 'text',
@@ -122,7 +130,7 @@ export const useAwsCredentialsForm = ({
       });
     } else {
       updatePolicy({
-        updatedPolicy: getPosturePolicy(newPolicy, awsPolicyType, {
+        updatedPolicy: updatePolicyWithInputs(newPolicy, awsPolicyType, {
           'aws.credentials.type': {
             // Restoring last manual credentials type or defaulting to the first option
             value: lastManualCredentialsType.current || DEFAULT_MANUAL_AWS_CREDENTIALS_TYPE,
@@ -154,7 +162,7 @@ const getAwsCloudFormationTemplate = (newPolicy: NewPackagePolicy, awsPolicyType
     return undefined;
   }
   const template: string | undefined = newPolicy?.inputs?.find((i) => i.type === awsPolicyType)
-    ?.vars?.cloud_formation_template_url?.value?.config?.cloud_formation_template_url?.value;
+    ?.config?.cloud_formation_template_url?.value;
 
   return template || undefined;
 };
@@ -166,11 +174,6 @@ const updateCloudFormationPolicyTemplate = (
   awsPolicyType?: string
 ) => {
   if (!awsPolicyType) {
-    return;
-  }
-  const stored = getAwsCloudFormationTemplate(newPolicy, awsPolicyType);
-
-  if (stored === templateUrl) {
     return;
   }
 
@@ -195,34 +198,30 @@ const useCloudFormationTemplate = ({
   newPolicy,
   updatePolicy,
   setupFormat,
+  awsPolicyType,
+  templateName,
 }: {
   packageInfo: PackageInfo;
   newPolicy: NewPackagePolicy;
   updatePolicy: UpdatePolicy;
   setupFormat: AwsSetupFormat;
+  awsPolicyType?: string;
+  templateName: string;
 }) => {
-  const { awsPolicyType, templateName } = useCloudSetup();
-  useEffect(() => {
-    const policyInputCloudFormationTemplate = getAwsCloudFormationTemplate(
-      newPolicy,
-      awsPolicyType
-    );
+  const policyInputCloudFormationTemplate = getAwsCloudFormationTemplate(newPolicy, awsPolicyType);
 
-    if (setupFormat === AWS_SETUP_FORMAT.MANUAL) {
-      if (policyInputCloudFormationTemplate) {
-        updateCloudFormationPolicyTemplate(newPolicy, updatePolicy, undefined, awsPolicyType);
-      }
-      return;
+  if (setupFormat === AWS_SETUP_FORMAT.MANUAL) {
+    if (policyInputCloudFormationTemplate) {
+      updateCloudFormationPolicyTemplate(newPolicy, updatePolicy, undefined, awsPolicyType);
     }
-    const templateUrl = getCloudFormationDefaultValue(packageInfo, templateName);
+    return;
+  }
+  const templateUrl = getCloudFormationDefaultValue(packageInfo, templateName);
 
-    // If the template is not available, do not update the policy
-    if (templateUrl === '') return;
+  // If the template is not available, do not update the policy
+  if (templateUrl === '') return;
 
-    // If the template is already set, do not update the policy
-    if (policyInputCloudFormationTemplate === templateUrl) return;
-
-    updateCloudFormationPolicyTemplate(newPolicy, updatePolicy, templateUrl, awsPolicyType);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [newPolicy?.vars?.cloud_formation_template_url, setupFormat]);
+  // If the template is already set, do not update the policy
+  if (policyInputCloudFormationTemplate === templateUrl) return;
+  updateCloudFormationPolicyTemplate(newPolicy, updatePolicy, templateUrl, awsPolicyType);
 };
