--- x-pack/solutions/security/packages/kbn-cloud-security-posture/public/src/components/fleet_extensions/aws_credentials_form/aws_utils.ts
@@ -1,107 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { NewPackagePolicyInput, PackageInfo } from '@kbn/fleet-plugin/common';
-import { SetupTechnology } from '@kbn/fleet-plugin/public';
-import {
-  DEFAULT_AGENTLESS_AWS_CREDENTIALS_TYPE,
-  DEFAULT_AGENTLESS_CLOUD_CONNECTORS_AWS_CREDENTIALS_TYPE,
-  DEFAULT_AWS_CREDENTIALS_TYPE,
-  DEFAULT_MANUAL_AWS_CREDENTIALS_TYPE,
-} from './get_aws_credentials_form_options';
-import { hasPolicyTemplateInputs } from '../utils';
-import { AWS_CREDENTIALS_TYPE } from '../constants';
-import { AwsCredentialsType } from '../types';
-
-export const getDefaultAwsCredentialsType = (
-  packageInfo: PackageInfo,
-  showCloudConnectors: boolean,
-  templateName: string,
-  setupTechnology?: SetupTechnology
-): AwsCredentialsType => {
-  if (setupTechnology && setupTechnology === SetupTechnology.AGENTLESS) {
-    return showCloudConnectors
-      ? AWS_CREDENTIALS_TYPE.CLOUD_CONNECTORS
-      : AWS_CREDENTIALS_TYPE.DIRECT_ACCESS_KEYS;
-  }
-  const hasCloudFormationTemplate = !!getCloudFormationDefaultValue(packageInfo, templateName);
-
-  if (hasCloudFormationTemplate) {
-    return AWS_CREDENTIALS_TYPE.CLOUD_FORMATION;
-  }
-
-  return DEFAULT_MANUAL_AWS_CREDENTIALS_TYPE;
-};
-
-export const getCloudDefaultAwsCredentialConfig = ({
-  isAgentless,
-  showCloudConnectors,
-  packageInfo,
-  templateName,
-}: {
-  isAgentless: boolean;
-  packageInfo: PackageInfo;
-  showCloudConnectors: boolean;
-  templateName: string;
-}) => {
-  let credentialsType;
-  const hasCloudFormationTemplate = !!getCloudFormationDefaultValue(packageInfo, templateName);
-  if (!showCloudConnectors && isAgentless) {
-    credentialsType = DEFAULT_AGENTLESS_AWS_CREDENTIALS_TYPE;
-  } else if (showCloudConnectors && isAgentless) {
-    credentialsType = DEFAULT_AGENTLESS_CLOUD_CONNECTORS_AWS_CREDENTIALS_TYPE;
-  } else if (hasCloudFormationTemplate && !isAgentless) {
-    credentialsType = DEFAULT_AWS_CREDENTIALS_TYPE;
-  } else {
-    credentialsType = DEFAULT_MANUAL_AWS_CREDENTIALS_TYPE;
-  }
-  const config: {
-    [key: string]: {
-      value: string | boolean;
-      type: 'text' | 'bool';
-    };
-  } = {
-    'aws.credentials.type': {
-      value: credentialsType,
-      type: 'text',
-    },
-    ...(showCloudConnectors && {
-      'aws.supports_cloud_connectors': {
-        value: showCloudConnectors,
-        type: 'bool',
-      },
-    }),
-  };
-
-  return config;
-};
-
-export const getAwsCredentialsType = (
-  input: NewPackagePolicyInput
-): AwsCredentialsType | undefined => input.streams[0].vars?.['aws.credentials.type'].value;
-
-export const getCloudFormationDefaultValue = (
-  packageInfo: PackageInfo,
-  templateName: string
-): string => {
-  if (!packageInfo.policy_templates) return '';
-
-  const policyTemplate = packageInfo.policy_templates.find((p) => p.name === templateName);
-  if (!policyTemplate) return '';
-
-  const policyTemplateInputs = hasPolicyTemplateInputs(policyTemplate) && policyTemplate.inputs;
-
-  if (!policyTemplateInputs) return '';
-
-  const cloudFormationTemplate = policyTemplateInputs.reduce((acc, input): string => {
-    if (!input.vars) return acc;
-    const template = input.vars.find((v) => v.name === 'cloud_formation_template')?.default;
-    return template ? String(template) : acc;
-  }, '');
-
-  return cloudFormationTemplate;
-};
