+++ x-pack/solutions/security/packages/kbn-cloud-security-posture/public/src/components/fleet_extensions/gcp_credentials_form/gcp_account_type_selector.tsx
@@ -13,12 +13,11 @@ import { NewPackagePolicyInput, PackageInfo } from '@kbn/fleet-plugin/common';
 import type { NewPackagePolicy } from '@kbn/fleet-plugin/public';
 import { EuiCallOut, EuiSpacer, EuiText } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { getPosturePolicy } from '../utils';
+import { updatePolicyWithInputs, gcpField, getGcpInputVarsFields } from '../utils';
 import { CspRadioGroupProps, RadioGroup } from '../csp_boxed_radio_group';
-import { gcpField, getInputVarsFields } from './gcp_utils';
 import { UpdatePolicy } from '../types';
 import { GCP_ORGANIZATION_ACCOUNT, GCP_SINGLE_ACCOUNT } from '../constants';
-import { useCloudSetup } from '../cloud_setup_context';
+import { useCloudSetup } from '../hooks/use_cloud_setup_context';
 
 const getGcpAccountTypeOptions = (isGcpOrgDisabled: boolean): CspRadioGroupProps['options'] => [
   {
@@ -86,7 +85,7 @@ export const GcpAccountTypeSelect = ({
   const subsetOfGcpField = (({ 'gcp.organization_id': a }) => ({ 'gcp.organization_id': a }))(
     gcpField.fields
   );
-  const fieldsToHide = getInputVarsFields(input, subsetOfGcpField);
+  const fieldsToHide = getGcpInputVarsFields(input, subsetOfGcpField);
   const fieldsSnapshot = useRef({});
   const lastSetupAccessType = useRef<string | undefined>(undefined);
   const onSetupFormatChange = (newSetupFormat: string) => {
@@ -98,7 +97,7 @@ export const GcpAccountTypeSelect = ({
       // We need to store the last manual credentials type to restore it later
       lastSetupAccessType.current = input.streams[0].vars?.['gcp.account_type'].value;
       updatePolicy({
-        updatedPolicy: getPosturePolicy(newPolicy, gcpPolicyType, {
+        updatedPolicy: updatePolicyWithInputs(newPolicy, gcpPolicyType, {
           'gcp.account_type': {
             value: 'single-account',
             type: 'text',
@@ -110,7 +109,7 @@ export const GcpAccountTypeSelect = ({
       });
     } else {
       updatePolicy({
-        updatedPolicy: getPosturePolicy(newPolicy, gcpPolicyType, {
+        updatedPolicy: updatePolicyWithInputs(newPolicy, gcpPolicyType, {
           'gcp.account_type': {
             // Restoring last manual credentials type
             value: lastSetupAccessType.current || 'organization-account',
@@ -126,7 +125,7 @@ export const GcpAccountTypeSelect = ({
   useEffect(() => {
     if (!getGcpAccountType(input)) {
       updatePolicy({
-        updatedPolicy: getPosturePolicy(newPolicy, gcpPolicyType, {
+        updatedPolicy: updatePolicyWithInputs(newPolicy, gcpPolicyType, {
           'gcp.account_type': {
             value: isGcpOrgDisabled ? GCP_SINGLE_ACCOUNT : GCP_ORGANIZATION_ACCOUNT,
             type: 'text',
