+++ x-pack/solutions/security/packages/kbn-cloud-security-posture/public/src/components/fleet_extensions/gcp_credentials_form/gcp_input_var_fields.tsx
@@ -17,9 +17,8 @@ import {
 import { LazyPackagePolicyInputVarField } from '@kbn/fleet-plugin/public';
 import { PackageInfo } from '@kbn/fleet-plugin/common';
 import { i18n } from '@kbn/i18n';
-import { fieldIsInvalid, findVariableDef } from '../utils';
+import { fieldIsInvalid, findVariableDef, gcpField } from '../utils';
 import { GCP_INPUT_FIELDS_TEST_SUBJECTS } from './gcp_test_subjects';
-import { gcpField } from './gcp_utils';
 import { GCP_CREDENTIALS_TYPE } from '../constants';
 import { GcpFields, GcpInputFields } from '../types';
