+++ x-pack/solutions/security/packages/kbn-cloud-security-posture/public/src/components/fleet_extensions/gcp_credentials_form/gcp_setup_info.tsx
@@ -7,7 +7,7 @@
 import React from 'react';
 import { EuiHorizontalRule, EuiLink, EuiSpacer, EuiText, EuiTitle } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { useCloudSetup } from '../cloud_setup_context';
+import { useCloudSetup } from '../hooks/use_cloud_setup_context';
 
 export const GCPSetupInfoContent = ({ isAgentless }: { isAgentless: boolean }) => {
   const { gcpOverviewPath } = useCloudSetup();
