+++ x-pack/solutions/security/test/cloud_security_posture_functional/pages/index.ts
@@ -19,26 +19,26 @@ export default function ({ getPageObjects, loadTestFile }: FtrProviderContext) {
       await cspSecurity.createUsers();
     });
 
-    loadTestFile(require.resolve('./rules/rules_counters'));
-    loadTestFile(require.resolve('./rules/rules_table'));
-    loadTestFile(require.resolve('./rules/rules_table_headers'));
-    loadTestFile(require.resolve('./findings_onboarding'));
-    loadTestFile(require.resolve('./findings'));
-    loadTestFile(require.resolve('./findings_grouping'));
-    loadTestFile(require.resolve('./findings_alerts'));
-    loadTestFile(require.resolve('./compliance_dashboard'));
-    loadTestFile(require.resolve('./vulnerability_dashboard'));
+    // loadTestFile(require.resolve('./rules/rules_counters'));
+    // loadTestFile(require.resolve('./rules/rules_table'));
+    // loadTestFile(require.resolve('./rules/rules_table_headers'));
+    // loadTestFile(require.resolve('./findings_onboarding'));
+    // loadTestFile(require.resolve('./findings'));
+    // loadTestFile(require.resolve('./findings_grouping'));
+    // loadTestFile(require.resolve('./findings_alerts'));
+    // loadTestFile(require.resolve('./compliance_dashboard'));
+    // loadTestFile(require.resolve('./vulnerability_dashboard'));
     loadTestFile(require.resolve('./cis_integrations/cnvm/cis_integration_cnvm'));
     loadTestFile(require.resolve('./cis_integrations/cspm/cis_integration_aws'));
     loadTestFile(require.resolve('./cis_integrations/cspm/cis_integration_gcp'));
     loadTestFile(require.resolve('./cis_integrations/cspm/cis_integration_azure'));
     loadTestFile(require.resolve('./cis_integrations/kspm/cis_integration_k8s'));
     loadTestFile(require.resolve('./cis_integrations/kspm/cis_integration_eks'));
-    loadTestFile(require.resolve('./findings_old_data'));
-    loadTestFile(require.resolve('./vulnerabilities'));
-    loadTestFile(require.resolve('./vulnerabilities_grouping'));
-    loadTestFile(require.resolve('./benchmark'));
-    loadTestFile(require.resolve('./alerts_flyout'));
-    loadTestFile(require.resolve('./events_flyout'));
+    // loadTestFile(require.resolve('./findings_old_data'));
+    // loadTestFile(require.resolve('./vulnerabilities'));
+    // loadTestFile(require.resolve('./vulnerabilities_grouping'));
+    // loadTestFile(require.resolve('./benchmark'));
+    // loadTestFile(require.resolve('./alerts_flyout'));
+    // loadTestFile(require.resolve('./events_flyout'));
   });
 }
